+++ Libs/Optimize/CMakeLists.txt
@@ -1,4 +1,4 @@
-FILE(GLOB Optimize_sources *.cpp)
+FILE(GLOB Optimize_sources *.cpp Domain/*.cpp Function/*.cpp)
 FILE(GLOB Optimize_headers *.h *.txx)
 
 set(SW_PYTHON_LIBS pybind11::embed)
