+++ Libs/Optimize/Domain/ImageDomainWithCurvature.h
