+++ Libs/Optimize/Domain/ImageDomainWithGradN.h
