+++ Libs/Optimize/Domain/ImageDomainWithGradients.h
