+++ Libs/Optimize/Function/DualVectorFunction.h
