+++ Libs/Optimize/Sampler.h
@@ -3,23 +3,23 @@
 #include <Logging.h>
 #include <Mesh/Mesh.h>
 
-#include "CorrespondenceFunction.h"
 #include "CorrespondenceMode.h"
-#include "CurvatureSamplingFunction.h"
 #include "DomainType.h"
-#include "DualVectorFunction.h"
 #include "GenericContainerArray.h"
 #include "GradientDescentOptimizer.h"
-#include "ImplicitSurfaceDomain.h"
-#include "LegacyCorrespondenceFunction.h"
+#include "Libs/Optimize/Domain/ImplicitSurfaceDomain.h"
+#include "Libs/Optimize/Domain/MeshDomain.h"
+#include "Libs/Optimize/Function/CorrespondenceFunction.h"
+#include "Libs/Optimize/Function/CurvatureSamplingFunction.h"
+#include "Libs/Optimize/Function/DualVectorFunction.h"
+#include "Libs/Optimize/Function/LegacyCorrespondenceFunction.h"
+#include "Libs/Optimize/Function/SamplingFunction.h"
 #include "MeanCurvatureContainer.h"
-#include "MeshDomain.h"
 #include "MeshWrapper.h"
 #include "ParticleShapeLinearRegressionMatrixAttribute.h"
 #include "ParticleShapeMixedEffectsMatrixAttribute.h"
 #include "ParticleSurfaceNeighborhood.h"
 #include "ParticleSystem.h"
-#include "SamplingFunction.h"
 #include "TriMesh.h"
 #include "vnl/vnl_matrix_fixed.h"
